Fixed unpredictable relative file path in gizmo.
The use of __file__ in knob is unpredictable since it'll return the last imported python file.
Moved __file__ to localizeLensKernel.py for it to be part of onCreate.
If nuke script is not saved, onCreate will now set absolute file paths for kernels.
If nuke script is saved, onCreate will copy the kernel files to the script folder, and set the knobs paths relative to script folder.